# build: Simplify WebpackRTLPlugin, use CssMinimizerPlugin for CSS minification