--- a/apps/wpcom-block-editor/package.json
+++ b/apps/wpcom-block-editor/package.json
@@ -50,7 +50,7 @@
 		"tinymce": "^4.9.6"
 	},
 	"devDependencies": {
-		"@automattic/calypso-build": "^7.0.0",
+		"@automattic/calypso-build": "^8.0.0",
 		"@wordpress/dependency-extraction-webpack-plugin": "^3.0.1",
 		"enzyme": "^3.11.0",
 		"npm-run-all": "^4.1.5"