--- a/packages/calypso-build/package.json
+++ b/packages/calypso-build/package.json
@@ -55,6 +55,7 @@
 		"browserslist": "^4.8.2",
 		"cache-loader": "^4.1.0",
 		"css-loader": "^3.4.2",
+		"css-minimizer-webpack-plugin": "^1.3.0",
 		"duplicate-package-checker-webpack-plugin": "^3.0.0",
 		"enzyme-adapter-react-16": "^1.15.1",
 		"enzyme-to-json": "^3.4.3",