--- a/packages/calypso-build/package.json
+++ b/packages/calypso-build/package.json
@@ -34,7 +34,7 @@
 		"extends @wordpress/browserslist-config"
 	],
 	"dependencies": {
-		"@automattic/webpack-rtl-plugin": "^4.0.0",
+		"@automattic/webpack-rtl-plugin": "^5.0.0",
 		"@babel/cli": "^7.13.10",
 		"@babel/compat-data": "^7.13.11",
 		"@babel/core": "^7.13.10",