--- a/packages/calypso-build/package.json
+++ b/packages/calypso-build/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@automattic/calypso-build",
-	"version": "7.0.0",
+	"version": "8.0.0",
 	"description": "Shared Calypso build configuration files",
 	"keywords": [
 		"babel",