--- a/packages/composite-checkout/package.json
+++ b/packages/composite-checkout/package.json
@@ -48,7 +48,7 @@
 		"react-stripe-elements": "^4.0.2"
 	},
 	"devDependencies": {
-		"@automattic/calypso-build": "^7.0.0",
+		"@automattic/calypso-build": "^8.0.0",
 		"@automattic/calypso-polyfills": "^1.0.0",
 		"@testing-library/jest-dom": "^5.9.0",
 		"@testing-library/react": "^10.0.5",