--- a/packages/whats-new/package.json
+++ b/packages/whats-new/package.json
@@ -44,7 +44,7 @@
 		"wpcom": "^6.0.0"
 	},
 	"devDependencies": {
-		"@automattic/calypso-build": "^7.0.0",
+		"@automattic/calypso-build": "^8.0.0",
 		"react": "^16.12.0"
 	},
 	"peerDependencies": {