### Background

We are using a plugin `@automattic/webpack-rtl-plugin` in Calypso builds that does two things: generate extra CSS files for RTL languages, and minimize all CSS files using `cssnano`. This plugin was originally forked from https://github.com/romainberger/webpack-rtl-plugin.

In this PR I've cleaned up the plugin code a bit, plus remove the capacity to minify CSS files. Instead, I'm using a new plugin `css-minimizer-webpack-plugin`. This new plugin also uses `cssnano`, but the main difference is that it supports parallelization, making the minification process faster.

Before (total duration: `40,287ms`)

```
#15 70.78 [evergreen] <s> [webpack.Progress] 92% sealing asset processing WebpackRTLPlugin
#15 111.1 [evergreen] <e> [webpack.Progress]  |  | 40287 ms asset processing > WebpackRTLPlugin
```

After (total duration: `11,160ms`):
```
#15 68.38 [evergreen] <s> [webpack.Progress] 92% sealing asset processing WebpackRTLPlugin
#15 71.33 [evergreen] <w> [webpack.Progress]  |  | 2946 ms asset processing > WebpackRTLPlugin
#15 85.09 [evergreen] <s> [webpack.Progress] 92% sealing asset processing CssMinimizerPlugin
#15 93.31 [evergreen] <w> [webpack.Progress]  |  | 8214 ms asset processing > CssMinimizerPlugin
```

It makes the build ~30s faster, while producing the same exact CSS files.


### Changes proposed in this Pull Request

* In `@automattic/webpack-rtl-plugin`:
  * Don't use `async` to process files in batches.
  * Modernize async code to use async/await promises.
  * Drop capacity to minify CSS code.
  * Drop capacity to define the filename for the RTS file. We don't use it and complicates the code a little bit.

* In `@automattic/calypso-build`:
  * Use `css-minimizer-webpack-plugin` to minimize code. 
 

### Testing instructions

* Load the live branch, switch to a RTL language and verify styles are correct.

* Compare CSS files:
  * Build this branch `rm -fr public/evergreen && CALYPSO_ENV=production NODE_ENV=production yarn build-client-evergreen`.
  * Do the same with `ci/webpack-profile` (the target branch of this PR)
  * Verify `public/evergreen/*.css` files are the same in both cases (`diff -qr wp-calypso/public/evergreen wp-calypso2/public/evergreen | grep css`)