@scinos 2021-03-26
> Can you just verify that none of the unsupported options are used in the major projects that consume calypso-build before shipping?

Good idea. Using opengrok I found that nobody except Calypso is using `calypso-build/webpack/minify` directly. All users of `calypso-build` just extend `calypso-build/webpack.config.js`.

But thanks to this search I found that Calypso _is_ passing `extractComments`, which I forgot to add to the new API 🤦🏼 . I'll fix it before merging.